[#176578322] rename CobadgeCard
