+++ ts/features/wallet/cobadge/component/BaseCoBadgeCard.tsx
